UI and perormance greatly improved
Download is now only generated once for each roster configuration, if not changed, it wont trigger another download generation. Also, thanks to downoloadjs, the dataurl for the roster is not appended to the html, improving performance by an egregious amount. Even at 4K resolutions, download is now responsive, so the warning was ditched.
As for UI, My Characters are now ordered by name, dropdowns cleaned up with proper headers, and delete my characters is now a dropdown item with custom stylings.